--- server/services/title.js
@@ -1,50 +1,0 @@
-const models = require("../models");
-
-function getTitle(req, res, next) {
-  models.users
-    .findOne({
-      where: { jwt: req.query.googleId },
-    })
-    .then((user) => {
-      var msg = {
-        success: true,
-        data: {
-          nickname: user.nickname,
-        },
-        message: "success",
-      };
-      res.send(msg);
-    })
-    .catch((err) => {
-      var msg = {
-        success: false,
-        message: err,
-      };
-      res.send(msg);
-    });
-}
-
-function setTitle(req, res, next) {
-  models.users
-    .update(
-      {
-        nickname: req.body.nickname,
-      },
-      { where: { jwt: req.body.googleId } }
-    )
-    .then(() => {
-      var temp = {
-        success: true,
-        message: "Success",
-      };
-      res.send(temp);
-    })
-    .catch((err) => {
-      console.error(err);
-    });
-}
-
-module.exports = {
-  setTitle,
-  getTitle,
-};
